requirements.txt
alembic==1.13.3
aniso8601==9.0.1
asttokens==2.4.1
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
databricks-sdk==0.33.0
debugpy==1.8.6
decorator==5.1.1
delta-spark==3.2.1
Deprecated==1.2.14
docker==7.1.0
executing==2.1.0
Flask==3.0.3
fonttools==4.54.1
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.35.0
graphene==3.3
graphql-core==3.2.4
graphql-relay==3.2.0
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
importlib_metadata==8.4.0
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.27.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
Mako==1.3.5
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mlflow==2.16.2
mlflow-skinny==2.16.2
nest-asyncio==1.6.0
numpy==2.1.1
nvidia-nccl-cu12==2.23.4
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
packaging==24.1
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
plotly==5.24.1
pluggy==1.5.0
prompt_toolkit==3.0.48
protobuf==5.28.2
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py4j==0.10.9.7
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
Pygments==2.18.0
pyparsing==3.1.4
pyspark==3.5.3
pytest==8.3.3
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
requests==2.32.3
rsa==4.9
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.35
sqlparse==0.5.1
stack-data==0.6.3
tenacity==9.0.0
threadpoolctl==3.5.0
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.0.4
wrapt==1.16.0
xgboost==2.1.1
zipp==3.20.2
-e .